Implement ERC20Permit functionality and refactor ERC20 and ERC20Storage code
This commit extends the `ERC20` contract to implement the `IERC20Permit` interface allowing for gasless transactions. It also refactors the ERC20Storage code to move the emissions of transfer and approval events from the `ERC20` contract to the `ERC20Storage`. Error handling has been improved in `ERC20Permit` where the signer of a permit is now validated.